# backend/docs/swagger.yaml
basePath: /
definitions:
...
  pgtype.InfinityModifier:
    enum:
    - 1
    - 0
    - -1
    type: integer
    x-enum-varnames:
    - Infinity
    - Finite
    - NegativeInfinity
...
  pgtype.Timestamp:
    properties:
      infinityModifier:
        $ref: '#/definitions/pgtype.InfinityModifier'
      time:
        description: Time zone will be ignored when encoding to PostgreSQL.
        type: string
      valid:
        type: boolean
    type: object
...
  server.CreateSymptomRequest:
    properties:
      symptom_description:
        type: string
      symptom_name:
        description: Changed to string (required)
        type: string
    type: object
...
  server.HTTPError:
    properties:
      error:
        description: Error message
        example: Resource not found
        type: string
    type: object
...
  server.SymptomResponse:
    properties:
      created_at:
        $ref: '#/definitions/pgtype.Timestamp'
      symptom_description:
        type: string
      symptom_id:
        type: integer
      symptom_name:
        description: Changed to string (cannot be null)
        type: string
      updated_at:
        $ref: '#/definitions/pgtype.Timestamp'
    type: object
...
  server.UpdateSymptomRequest:
    properties:
      symptom_description:
        type: string
      symptom_name:
        description: Changed to string (required)
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API Service for managing patients, symptoms, and diseases.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Patient API Service
  version: "1.0"
paths:
...
  /symptoms:
    get:
      consumes:
      - application/json
      description: Get a list of all symptoms
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved list of symptoms
          schema:
            items:
              $ref: '#/definitions/server.SymptomResponse'
            type: array
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/server.HTTPError'
      summary: List symptoms
      tags:
      - Symptoms
    post:
      consumes:
      - application/json
      description: Add a new symptom record. Symptom name is required.
      parameters:
      - description: Symptom data to create
        in: body
        name: symptom
        required: true
        schema:
          $ref: '#/definitions/server.CreateSymptomRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Symptom created successfully
          schema:
            $ref: '#/definitions/server.SymptomResponse'
        "400":
          description: Invalid request payload or missing symptom name
          schema:
            $ref: '#/definitions/server.HTTPError'
        "409":
          description: Symptom name already exists (unique constraint)
          schema:
            $ref: '#/definitions/server.HTTPError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/server.HTTPError'
      summary: Create a new symptom
      tags:
      - Symptoms
  /symptoms/{symptomID}:
    delete:
      consumes:
      - application/json
      description: Delete a symptom record by ID. Associated records in junction tables
        (patient_symptoms, patient_disease_symptom) should be removed via ON DELETE
        CASCADE.
      parameters:
      - description: Symptom ID
        format: int32
        in: path
        name: symptomID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content (Successful deletion)
          schema:
            type: string
        "400":
          description: Invalid Symptom ID format
          schema:
            $ref: '#/definitions/server.HTTPError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/server.HTTPError'
      summary: Delete symptom
      tags:
      - Symptoms
    get:
      consumes:
      - application/json
      description: Retrieve details of a specific symptom by its ID
      parameters:
      - description: Symptom ID
        format: int32
        in: path
        name: symptomID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved symptom
          schema:
            $ref: '#/definitions/server.SymptomResponse'
        "400":
          description: Invalid Symptom ID format
          schema:
            $ref: '#/definitions/server.HTTPError'
        "404":
          description: Symptom not found
          schema:
            $ref: '#/definitions/server.HTTPError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/server.HTTPError'
      summary: Get symptom by ID
      tags:
      - Symptoms
    put:
      consumes:
      - application/json
      description: Update details for an existing symptom. Symptom name is required.
      parameters:
      - description: Symptom ID
        format: int32
        in: path
        name: symptomID
        required: true
        type: integer
      - description: Symptom data to update
        in: body
        name: symptom
        required: true
        schema:
          $ref: '#/definitions/server.UpdateSymptomRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Symptom updated successfully
          schema:
            $ref: '#/definitions/server.SymptomResponse'
        "400":
          description: Invalid request payload or missing symptom name
          schema:
            $ref: '#/definitions/server.HTTPError'
        "404":
          description: Symptom not found
          schema:
            $ref: '#/definitions/server.HTTPError'
        "409":
          description: Symptom name already exists (unique constraint)
          schema:
            $ref: '#/definitions/server.HTTPError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/server.HTTPError'
      summary: Update symptom details
      tags:
      - Symptoms
schemes:
- http
- https
swagger: "2.0"